atvs 19 and 20

+++ Caminhao.java
@@ -1,8 +1,8 @@
 public class Caminhao extends Veiculo {
-    private int capacidade;
+    private double capacidade;
 	private int numEixos;
 	
-	public Caminhao(String modelo, String placa, int anoFabr, int pCapacidade, int pNumEixos, double valor) {
+	public Caminhao(String modelo, String placa, int anoFabr, double pCapacidade, int pNumEixos, double valor) {
      	super(modelo, placa, anoFabr, valor);
 		setCapacidade(pCapacidade);
 		setNumeroEixos(pNumEixos);
@@ -18,16 +18,16 @@ public class Caminhao extends Veiculo {
 		return numEixos;
 	}
 
-    public void setCapacidade(int capacidade) {
+    public void setCapacidade(double capacidade) {
 		this.capacidade = capacidade;
 	}
 
-    public int getCapacidade() {
+    public double getCapacidade() {
 		return capacidade;
 	}
 
     public void imprime(){
 		System.out.printf("\nVeiculo: %s\nPlaca: %7s\nFabr: %4d", modelo, placa, anoFabr);
-        System.out.printf("\nCapacidade: %d\nEixos: %d\nR$ %.2f\n", capacidade, numEixos, valor);
+        System.out.printf("\nCapacidade: %.2f\nEixos: %d\nR$ %.2f\n", capacidade, numEixos, valor);
     }
 }
